+++ CHANGELOG.md
@@ -2,9 +2,17 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.2] - 2015-04-23
+### Changed
+* Fixed the timerSet callback to return null
+
+### Added 
+* added a type property on the cache 
+
 ## [0.0.1] - 2015-04-23
 ### Added
 * Support for insert, select, counting, and removing data
 * 12 tests passing 
 
+[0.0.2]: https://github.com/Esri/koop-pgcache/tags/v0.0.2
 [0.0.1]: https://github.com/Esri/koop-pgcache/tags/v0.0.1